+++ src/extension.ts
@@ -5,14 +5,25 @@ import {
   commands,
   ExtensionContext,
   FileSystemWatcher,
-  WorkspaceConfiguration
+  GlobPattern,
 } from 'vscode'
+
+type ConfigProperties = {
+  monitorFilesForTypescript: boolean
+  monitorFilesForESLint: boolean
+  fileGlobForTypescript: GlobPattern
+  fileGlobForESLint: GlobPattern
+  showRestartNotificationForTypescript: boolean
+  showRestartNotificationForESLint: boolean
+}
 
 const TS_EXT_ID = 'vscode.typescript-language-features'
 const ESLINT_EXT_ID = 'dbaeumer.vscode-eslint'
 const THIS_EXT_NAME = 'vscode-auto-restart-typescript-eslint-servers'
 const THIS_EXT_ID = `neotan.${THIS_EXT_NAME}`
+const THIS_EXT_CONFIG_PREFIX = `autoRestart` // i.e. Configuration `section`
 
+// Message templates
 const TS_SERVER_RESTARTED_MSG = `TypeScript Server Restarted as file(s) `
 const TS_SERVER_RESTART_FAILED_MSG =
   `TypeScript Server Restart failed when the file `
@@ -24,61 +35,46 @@ let tsWatcher: FileSystemWatcher
 let eslintWatcher: FileSystemWatcher
 
 export function activate(context: ExtensionContext) {
-  const wsConfig = workspace.getConfiguration('autoRestartTsESLint')
-  console.warn(wsConfig)
+  workspace.onDidChangeConfiguration((e) => {
 
-  if (wsConfig.monitorFilesForTypescript) {
-    tsWatcher = initWatcherForTypescriptFiles(wsConfig)
+    // Re-initiate the watchers might be overkill when any configuration 
+    // changed, but it's the easiest way to make sure the watchers are 
+    // up-to-date with the latest configuration.
+    if (e.affectsConfiguration(THIS_EXT_CONFIG_PREFIX)) {
+      tsWatcher?.dispose()
+      eslintWatcher?.dispose()
+
+      if (getConfig('monitorFilesForTypescript')) {
+        tsWatcher = initWatcherForTypescript()
+      }
+
+      if (getConfig('monitorFilesForESLint')) {
+        eslintWatcher = initWatcherForEslint()
+      }
+    }
+  })
+
+  if (getConfig('monitorFilesForTypescript')) {
+    tsWatcher = initWatcherForTypescript()
   }
 
-  if (wsConfig.monitorFilesForESLint) {
-    eslintWatcher = initWatcherForEslintFiles(wsConfig)
+  if (getConfig('monitorFilesForESLint')) {
+    eslintWatcher = initWatcherForEslint()
   }
-
-  console.log(
-    `Extension ${THIS_EXT_ID} is now active!`,
-    JSON.stringify(wsConfig, [
-      'monitorFilesForTypescript',
-      'monitorFilesForESLint',
-      'includeFilesForTypescript',
-      'includeFilesForESLint',
-    ], 2))
-
-  context.subscriptions.push(
-    commands
-      .registerCommand(`${THIS_EXT_NAME}.enableMonitorFilesForTypescript`,
-        async () => {
-          await wsConfig.update('monitorFilesForTypescript', true)
-          tsWatcher = initWatcherForTypescriptFiles(wsConfig)
-        }),
-    commands
-      .registerCommand(`${THIS_EXT_NAME}.disableMonitorFilesForTypescript`,
-        async () => {
-          await wsConfig.update('monitorFilesForTypescript', false)
-          tsWatcher?.dispose()
-        }),
-    commands
-      .registerCommand(`${THIS_EXT_NAME}.enableMonitorFilesForESLint`,
-        async () => {
-          await wsConfig.update('monitorFilesForESLint', true)
-          eslintWatcher = initWatcherForEslintFiles(wsConfig)
-        }),
-    commands
-      .registerCommand(`${THIS_EXT_NAME}.disableMonitorFilesForESLint`,
-        async () => {
-          await wsConfig.update('monitorFilesForESLint', false)
-          eslintWatcher?.dispose()
-        }),
-  )
 }
 
 export function deactivate() {
   tsWatcher?.dispose()
   eslintWatcher?.dispose()
-  console.log(`Extension ${THIS_EXT_ID} is now deactive!`)
+  console.log(`Extension ${THIS_EXT_ID} is now deactivated!`)
 }
 
 // ===== Utils =====
+
+function getConfig<K extends keyof ConfigProperties>(
+  property: K): ConfigProperties[K] {
+  return workspace.getConfiguration(THIS_EXT_CONFIG_PREFIX).get(property)!
+}
 
 function restartTsServer() {
   const tsExtension = extensions.getExtension(TS_EXT_ID)
@@ -100,10 +96,12 @@ function restartEslintServer() {
   return commands.executeCommand("eslint.restart")
 }
 
-function initWatcherForTypescriptFiles(wsConfig: WorkspaceConfiguration) {
+function initWatcherForTypescript(
+  fileGlobForTypescript: GlobPattern = getConfig('fileGlobForTypescript')
+) {
   const tsWatcher: FileSystemWatcher =
     workspace.createFileSystemWatcher(
-      wsConfig.includeFilesForTypescript,
+      fileGlobForTypescript,
       false,
       false,
       false
@@ -113,12 +111,14 @@ function initWatcherForTypescriptFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartTsServer()
-      window.showInformationMessage(
-        `${TS_SERVER_RESTARTED_MSG} was created: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForTypescript')) {
+        window.showInformationMessage(
+          `${TS_SERVER_RESTARTED_MSG} created: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${TS_SERVER_RESTART_FAILED_MSG} was created: 
+        `${TS_SERVER_RESTART_FAILED_MSG} created: 
           "${filePath}"`,
         { cause: err }
       )
@@ -129,12 +129,14 @@ function initWatcherForTypescriptFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartTsServer()
-      window.showInformationMessage(
-        `${TS_SERVER_RESTARTED_MSG} was changed: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForTypescript')) {
+        window.showInformationMessage(
+          `${TS_SERVER_RESTARTED_MSG} changed: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${TS_SERVER_RESTART_FAILED_MSG} was changed: 
+        `${TS_SERVER_RESTART_FAILED_MSG} changed: 
           "${filePath}"`,
         { cause: err }
       )
@@ -145,12 +147,14 @@ function initWatcherForTypescriptFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartTsServer()
-      window.showInformationMessage(
-        `${TS_SERVER_RESTARTED_MSG} was deleted: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForTypescript')) {
+        window.showInformationMessage(
+          `${TS_SERVER_RESTARTED_MSG} deleted: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${TS_SERVER_RESTART_FAILED_MSG} was deleted: 
+        `${TS_SERVER_RESTART_FAILED_MSG} deleted: 
           "${filePath}"`,
         { cause: err }
       )
@@ -160,10 +164,12 @@ function initWatcherForTypescriptFiles(wsConfig: WorkspaceConfiguration) {
   return tsWatcher
 }
 
-function initWatcherForEslintFiles(wsConfig: WorkspaceConfiguration) {
+function initWatcherForEslint(
+  fileGlobForESLint: GlobPattern = getConfig('fileGlobForESLint')
+) {
   const eslintWatcher: FileSystemWatcher =
     workspace.createFileSystemWatcher(
-      wsConfig.includeFilesForESLint,
+      fileGlobForESLint,
       false,
       false,
       false
@@ -173,12 +179,14 @@ function initWatcherForEslintFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartEslintServer()
-      window.showInformationMessage(
-        `${ESLINT_SERVER_RESTARTED_MSG} was created: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForESLint')) {
+        window.showInformationMessage(
+          `${ESLINT_SERVER_RESTARTED_MSG} created: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${ESLINT_SERVER_RESTART_FAILED_MSG} was Created: 
+        `${ESLINT_SERVER_RESTART_FAILED_MSG} created: 
           "${filePath}"`,
         { cause: err }
       )
@@ -189,12 +197,14 @@ function initWatcherForEslintFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartEslintServer()
-      window.showInformationMessage(
-        `${ESLINT_SERVER_RESTARTED_MSG} was changed: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForESLint')) {
+        window.showInformationMessage(
+          `${ESLINT_SERVER_RESTARTED_MSG} changed: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${ESLINT_SERVER_RESTART_FAILED_MSG} was changed: 
+        `${ESLINT_SERVER_RESTART_FAILED_MSG} changed: 
           "${filePath}"`,
         { cause: err }
       )
@@ -205,12 +215,14 @@ function initWatcherForEslintFiles(wsConfig: WorkspaceConfiguration) {
     const filePath = e?.path || e?.fsPath
     try {
       await restartEslintServer()
-      window.showInformationMessage(
-        `${ESLINT_SERVER_RESTARTED_MSG} was deleted: ${filePath}`
-      )
+      if (getConfig('showRestartNotificationForESLint')) {
+        window.showInformationMessage(
+          `${ESLINT_SERVER_RESTARTED_MSG} deleted: ${filePath}`
+        )
+      }
     } catch (err) {
       throw new Error(
-        `${ESLINT_SERVER_RESTART_FAILED_MSG} was deleted: 
+        `${ESLINT_SERVER_RESTART_FAILED_MSG} deleted: 
           "${filePath}"`,
         { cause: err }
       )