+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+### 0.0.5 (2023-07-20)
+* Add notification toggles when server restart (#1 & #5)
+* Fix files monitoring bug (#6)
+
 ### 0.0.4 (2023-07-18)
-* fix .eslintrc.js typo
-* add auto publishing GitHub Action
+* Fix `eslintrc` glob pattern typo ([@kentcdodds](https://github.com/kentcdodds) in #2)
+* Add auto publishing GitHub Action
 
 ### 0.0.3 (2023-01-30)
 Update icon